Move device connection dialog to its own widget
