+++ lib/main.dart
@@ -91,6 +91,9 @@ part 'ui/pages/camera_page.dart';
 part 'ui/pages/devices_page.dart';
 part 'ui/pages/loading_page.dart';
 part 'ui/pages/error_page.dart';
+part 'ui/pages/cans_login_android_webview.dart';
+part 'ui/pages/cans_login.dart';
+part 'ui/pages/connection_dialog.dart';
 
 part 'ui/widgets/study_card.dart';
 part 'ui/widgets/horizontal_bar.dart';
@@ -107,8 +110,6 @@ part 'ui/cards/survey_card.dart';
 part 'ui/cards/media_card.dart';
 part 'ui/cards/scoreboard_card.dart';
 part 'ui/cards/study_progress_card.dart';
-part 'ui/pages/cans_login_android_webview.dart';
-part 'ui/pages/cans_login.dart';
 part 'ui/cards/activity_card.dart';
 
 part 'main.g.dart';
